Properly implement custom fan modes
Doing this was a bit tricky, since there seems to be three custom modes:
Deep control, auto-maximum and fixed speed. Deep control mode is known
as "custom mode" in the code because "deep control" allows for a custom
fan curve.

Both auto-maximum mode and fixed speed mode require custom mode enabled
to function. Auto-maximum mode was the tricky part, as there is no
method that returns the value of the bit controlling it. It also has no
method of its own that can disable it, but the gaming mode and silent
mode methods can do that.

A helper method has been created to explicitly disable custom fan mode.
The helper method to set the fan mode is already large enough as it is
with all of the if-else statements it has to do.

While we are at it, fix some compilation errors and other weird bugs.