Add DMI checking to module
This code should not have made it into the driver, as it was not ready.
Since it did anyway, it might as well be used.

Currently, this will restrict the driver to only running on Aero or
AORUS laptops. Adding support for Gigabyte Gaming laptops will come
later, if it is necessary.