Save raw value for custom fan speed
The embedded controller recognizes 100 percent as 0xFF, but we are
sending our converted value of 0x64. This is causing it to set a new
custom fan speed when it is not supposed to while switching to
auto-maximum mode.

Fix this by saving the raw value during probe and when setting a new
custom fan speed.