Rename kernel driver to aorus-laptop
This driver is unlikely to work with Gigabyte Gaming models anyway, and
I would rather not add more code to it to explicitly support those
models. That would require a massive rewrite of the kernel driver, which
will include migrating to the newer device-based WMI API from the
deprecated GUID-based one we are using now.

For this, just rename this driver to aorus-laptop. A new kernel driver
will be developed that will only support Gigabyte Gaming laptops.