Implement probing fan curve
Gigabyte supports custom fan curves for custom fan mode, but the way
they implemented it requires calling WMI multiple times to get all of
the points. That could be why on Windows, there is a noticeable lag when
switching to fan control on their control center.

Right now, setting a custom fan curve is not supported yet, and adding
support for it is likely to result in a bit of ugliness. Even just to
get the necessary data, the goal should be to minimize calls to WMI
methods and avoid making it a hassle for the user to get the data they
need.